@@ -126,23 +126,23 @@ namespace Bannerlord.ButterLib.Implementation.CampaignIdentifier
         private object GetDescriptorInfoPiece(DescriptorAttribute descriptorAttribute) => descriptorAttribute switch
         {
             DescriptorAttribute.HeroName => FieldAccessHelper.TextObjectValueByRef(_baseHero.Name),
-            DescriptorAttribute.HeroFamilyName => _baseHero.Clan != null
+            DescriptorAttribute.HeroFamilyName => _baseHero.Clan is not null
                 ? FieldAccessHelper.TextObjectValueByRef(_baseHero.Clan.Name)
                 : string.Empty,
             DescriptorAttribute.HeroAge => (int) _baseHero.Age,
             DescriptorAttribute.HeroGender => _baseHero.IsFemale
                 ? 1
                 : 0,
-            DescriptorAttribute.HeroCulture => _baseHero.Culture != null
+            DescriptorAttribute.HeroCulture => _baseHero.Culture is not null
                 ? FieldAccessHelper.TextObjectValueByRef(_baseHero.Culture.Name)
                 : string.Empty,
-            DescriptorAttribute.FatherName => _baseHero.Father != null
+            DescriptorAttribute.FatherName => _baseHero.Father is not null
                 ? FieldAccessHelper.TextObjectValueByRef(_baseHero.Father.Name)
                 : string.Empty,
-            DescriptorAttribute.MotherName => _baseHero.Mother != null
+            DescriptorAttribute.MotherName => _baseHero.Mother is not null
                 ? FieldAccessHelper.TextObjectValueByRef(_baseHero.Mother.Name)
                 : string.Empty,
-            DescriptorAttribute.BirthplaceName => _baseHero.BornSettlement != null
+            DescriptorAttribute.BirthplaceName => _baseHero.BornSettlement is not null
                 ? FieldAccessHelper.TextObjectValueByRef(_baseHero.BornSettlement.Name)
                 : string.Empty,
             DescriptorAttribute.CharacterCode => CharacterCode.CreateFrom(_baseHero.CharacterObject).Code,
