@@ -58,7 +58,7 @@ namespace Bannerlord.ButterLib
             }
 
             var gameVersion = ApplicationVersionUtils.TryParse(ApplicationVersionUtils.GameVersionStr(), out var v) ? v : (ApplicationVersion?)null;
-            if (gameVersion == null)
+            if (gameVersion is null)
             {
                 logger?.LogError("Failed to get Game version!");
                 yield break;
@@ -132,14 +132,14 @@ namespace Bannerlord.ButterLib
             foreach (var subModuleType in subModules)
             {
                 var constructor = subModuleType.GetConstructor(BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.CreateInstance, null, Type.EmptyTypes, null);
-                if (constructor == null)
+                if (constructor is null)
                 {
                     logger?.LogError("SubModule {subModuleType} is missing a default constructor!", subModuleType);
                     continue;
                 }
 
                 var constructorFunc = AccessTools2.GetDelegate<ConstructorDelegate>(constructor);
-                if (constructorFunc == null)
+                if (constructorFunc is null)
                 {
                     logger?.LogError("SubModule {subModuleType}'s default constructor could not be converted to a delegate!", subModuleType);
                     continue;
@@ -155,9 +155,9 @@ namespace Bannerlord.ButterLib
         {
             using var assemblyVerifier = new AssemblyVerifier("ButterLib");
             var assemblyLoader = assemblyVerifier.GetLoader(out var exception);
-            if (assemblyLoader == null)
+            if (assemblyLoader is null)
             {
-                if (exception != null)
+                if (exception is not null)
                     logger?.LogError(0, exception, "AssemblyLoadProxy could not be initialized.");
                 else
                     logger?.LogError("AssemblyLoadProxy could not be initialized.");
